Cleaning up most codesmells
