+++ pkg/database/gorm/user/schema.go
@@ -7,6 +7,7 @@ import (
 	"github.com/jinzhu/gorm"
 )
 
+// User struct
 type User struct {
 	ID          uuid.UUID `gorm:"type:uuid;primary_key;"`
 	Email       string    `gorm:"unique;index;"`
@@ -17,6 +18,7 @@ type User struct {
 	ActivatedAt *time.Time
 }
 
+// BeforeCreate executes a hook before create the database entry
 func (user User) BeforeCreate(scope *gorm.Scope) error {
 	id := uuid.New()
 	err := scope.SetColumn("ID", id)
