+++ pkg/transport/grpc/server.go
@@ -1,7 +1,11 @@
 package grpc
 
 import (
+	"context"
+
 	"github.com/go-kit/kit/transport/grpc"
+	"github.com/sumelms/microservice-account/pkg/endpoint/user"
+	protouser "github.com/sumelms/microservice-account/proto/user"
 )
 
 type server struct {
@@ -10,4 +14,30 @@ type server struct {
 	updateUser grpc.Handler
 	deleteUser grpc.Handler
 	listUsers  grpc.Handler
+}
+
+// NewGRPCServer reates grpc server
+func NewGRPCServer(_ context.Context, endpoints user.Endpoints) protouser.UserServer {
+	return &server{
+		createUser: grpc.NewServer(
+			endpoints.CreateUser,
+			decodeCreateUserRequest,
+			encodeCreateUserResponse),
+		getUser: grpc.NewServer(
+			endpoints.GetUser,
+			decodeGetUserRequest,
+			encodeGetUserResponse),
+		updateUser: grpc.NewServer(
+			endpoints.UpdateUser,
+			decodeUpdateUserRequest,
+			encodeUpdateUserResponse),
+		deleteUser: grpc.NewServer(
+			endpoints.DeleteUser,
+			decodeDeleteUserRequest,
+			encodeDeleteUserResponse),
+		listUsers: grpc.NewServer(
+			endpoints.ListUsers,
+			decodeListUsersRequest,
+			encodeListUsersResponse),
+	}
 }
